+++ test/src/instance/sale/tef_ip_sale_payment_test.dart
@@ -10,6 +10,7 @@ import 'package:test/test.dart';
 
 import '../../../../testing/mocks/models/sale_mutation_response_model_mock.dart';
 import '../../../../testing/mocks/models/sale_payment_model_mock.dart';
+import '../../../../testing/mocks/models/sale_coupon_model_mock.dart';
 import '../../../../testing/mocks/networking/tef_ip_networking_client_test.dart';
 import '../../../../testing/mocks/shared/uri_mock.dart';
 
@@ -300,6 +301,78 @@ void main() {
             paymentId: kSalePaymentId,
             client: kHttpClient,
           ),
+          throwsA(isA<TefIPUnexpectedException>()),
+        );
+      });
+    });
+
+    group('clear', () {
+      test('should return SaleCouponModel on success', () async {
+        final expectedUrl =
+            TefIpUrlBuilder.build(TefIPEndpoints.salePaymentClear);
+
+        when(
+          () => kHttpClient.delete(
+            any(),
+            headers: any(named: 'headers'),
+          ),
+        ).thenAnswer(
+          (_) async => http.Response.bytes(
+            utf8.encode(jsonEncode(kSaleCoupon.toJson())),
+            200,
+          ),
+        );
+
+        final result = await salePayment.clear(client: kHttpClient);
+
+        expect(result, equals(kSaleCoupon));
+
+        verify(
+          () => kHttpClient.delete(
+            Uri.parse(expectedUrl),
+            headers: any(named: 'headers'),
+          ),
+        ).called(1);
+      });
+
+      test('should throw TefIPRequestException on ClientException', () {
+        when(
+          () => kHttpClient.delete(
+            any(),
+            headers: any(named: 'headers'),
+          ),
+        ).thenThrow(httpError);
+
+        expect(
+          () => salePayment.clear(client: kHttpClient),
+          throwsA(isA<TefIPRequestException>()),
+        );
+      });
+
+      test('should rethrow TefIPRequestException', () {
+        when(
+          () => kHttpClient.delete(
+            any(),
+            headers: any(named: 'headers'),
+          ),
+        ).thenThrow(TefIPRequestException(message: 'fail', statusCode: 400));
+
+        expect(
+          () => salePayment.clear(client: kHttpClient),
+          throwsA(isA<TefIPRequestException>()),
+        );
+      });
+
+      test('should throw TefIPUnexpectedException on unknown error', () {
+        when(
+          () => kHttpClient.delete(
+            any(),
+            headers: any(named: 'headers'),
+          ),
+        ).thenThrow(Exception());
+
+        expect(
+          () => salePayment.clear(client: kHttpClient),
           throwsA(isA<TefIPUnexpectedException>()),
         );
       });