+++ testing/mocks/models/sale_addition_model_mock.dart
@@ -1,0 +1,7 @@
+import 'package:dart_tefip/src/core/models/export.dart';
+
+final mockSaleAdditionModel = SaleAdditionModel(
+  id: 'ACR-001',
+  description: 'Acréscimo mock',
+  value: 5.0,
+);