+++ testing/mocks/models/sale_discount_model_mock.dart
@@ -1,0 +1,7 @@
+import 'package:dart_tefip/src/core/models/export.dart';
+
+final mockSaleDiscountModel = SaleDiscountModel(
+  id: 'DESC-001',
+  description: 'Desconto mock',
+  value: 10.0,
+);